move Landmark class to landmark_detector.py
